+++ packages/react-api/src/hooks/useCartoLayerProps.js
@@ -2,13 +2,11 @@ import { DataFilterExtension } from '@deck.gl/extensions';
 import { debounce, _buildFeatureFilter } from '@carto/react-core';
 import useViewportFeatures from './useViewportFeatures';
 
-export default function useCartoLayerProps(
-  source,
-  uniqueIdProperty = 'cartodb_id'
-) {
+export default function useCartoLayerProps(source, uniqueIdProperty) {
   const [onViewportLoad] = useViewportFeatures(source, uniqueIdProperty);
 
   return {
+    binary: true,
     uniqueIdProperty,
     onViewportLoad: debounce(onViewportLoad),
     getFilterValue: _buildFeatureFilter({ filters: source?.filters, type: 'number' }),