+++ packages/react-widgets/src/widgets/CategoryWidget.js
@@ -32,7 +32,7 @@ function CategoryWidget(props) {
   const [selectedCategories, setSelectedCategories] = useState([]);
   const dispatch = useDispatch();
   const source = useSelector((state) => selectSourceById(state, props.dataSource) || {});
-  const viewportFeatures = useSelector((state) => state.carto.viewportFeatures);
+  const viewportFeaturesReady = useSelector((state) => state.carto.viewportFeaturesReady);
   const widgetsLoadingState = useSelector((state) => state.carto.widgetsLoadingState);
   const [hasLoadingState, setIsLoading] = useWidgetLoadingState(
     props.id,
@@ -50,7 +50,8 @@ function CategoryWidget(props) {
         data,
         filters,
         credentials,
-        viewportFeatures: viewportFeatures[props.dataSource] || [],
+        viewportFeatures: viewportFeaturesReady[props.dataSource] || false,
+        dataSource: props.dataSource,
         type,
         opts: { abortController }
       })
@@ -73,7 +74,7 @@ function CategoryWidget(props) {
     setIsLoading,
     source.filters,
     type,
-    viewportFeatures,
+    viewportFeaturesReady,
     props,
     hasLoadingState
   ]);