+++ packages/react-widgets/src/widgets/HistogramWidget.js
@@ -3,8 +3,8 @@ import { useSelector, useDispatch } from 'react-redux';
 import { PropTypes } from 'prop-types';
 import { addFilter, removeFilter, selectSourceById } from '@carto/react-redux';
 import { WrapperWidgetUI, HistogramWidgetUI } from '@carto/react-ui';
-import { 
-  _FilterTypes as FilterTypes, 
+import {
+  _FilterTypes as FilterTypes,
   _getApplicableFilters as getApplicableFilters,
   AggregationTypes
 } from '@carto/react-core';
@@ -33,7 +33,7 @@ function HistogramWidget(props) {
   const [selectedBars, setSelectedBars] = useState([]);
   const dispatch = useDispatch();
   const source = useSelector((state) => selectSourceById(state, props.dataSource) || {});
-  const viewportFeatures = useSelector((state) => state.carto.viewportFeatures);
+  const viewportFeaturesReady = useSelector((state) => state.carto.viewportFeaturesReady);
   const widgetsLoadingState = useSelector((state) => state.carto.widgetsLoadingState);
   const [hasLoadingState, setIsLoading] = useWidgetLoadingState(
     props.id,
@@ -68,7 +68,8 @@ function HistogramWidget(props) {
         data,
         filters,
         credentials,
-        viewportFeatures: viewportFeatures[props.dataSource] || [],
+        viewportFeatures: viewportFeaturesReady[props.dataSource] || false,
+        dataSource: props.dataSource,
         type,
         opts: { abortController }
       })
@@ -91,7 +92,7 @@ function HistogramWidget(props) {
     setIsLoading,
     source.filters,
     type,
-    viewportFeatures,
+    viewportFeaturesReady,
     props,
     hasLoadingState
   ]);