+++ webpack.base.js
@@ -16,7 +16,7 @@ module.exports = {
     rules: [
       {
         test: /\.jsx?$/,
-        exclude: /(node_modules)/,
+        exclude: [/(node_modules)/, /(dist)/],
         loader: 'babel-loader'
       }
     ]